exit on error.

+++ freezing/build.sh
@@ -5,6 +5,10 @@ PYTHONBIN=$PYTHON/bin/python3.3
 export PYTHONPATH=..
 cd ..
 $PYTHONBIN setup.py install
+if [ $? -ne 0 ]; then
+    exit 1
+fi
+    
 cd freezing
 
 rm -rf dist
@@ -13,6 +17,10 @@ rm -rf build
 mkdir dist
 
 $PYTHONBIN $PYTHON/Tools/freeze/freeze.py -m  -o build kaa_freeze.py kaa.filetype.html kaa.filetype.html.htmlmode  kaa.filetype.javascript kaa.filetype.javascript.javascriptmode  kaa.filetype.css kaa.filetype.css.cssmode  kaa.filetype.python kaa.filetype.python.pythonmode  
+
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 
 cd build
 make
